requirements.txt
pytest==8.3.4
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
Jinja2==3.1.4
PyYAML==6.0.2
supabase==2.10.0
python-multipart==0.0.9
firebase-admin==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywebpush==1.14.0
cryptography==45.0.6
langchain-core==0.3.76
langgraph==0.2.69
dateparser==1.2.1
paramiko==3.5.1
websocket-client==1.8.0
